硬件数据数据json落盘：stars_soc和stars_chip_trans
# message auto-generated for no-merge-commit merge:
merge develop into master

硬件数据数据json落盘：stars_soc和stars_chip_trans

Created-by: msprofDeveloper

Match-id-50c9dd0fcae6c8d7b00df1aa545c78403b155644

+++ analysis/csrc/domain/entities/json_trace/trace_event.cpp
@@ -62,7 +62,7 @@ void CounterEvent::ProcessArgs(JsonWriter &ostream)
     }
 }
 
-void CounterEvent::SetSeriesValue(const std::string &key, const uint64_t &value)
+void CounterEvent::SetSeriesValue(const std::string &key, const double &value)
 {
     seriesValue_[key] = value;
 }
